Atmel Start - 27 Aug 19:17 UTC Last of these from Atmel Start.
This audit trail in git should be a good representation of
the Atmel Start generated USART project for SAMD51 -- good
enough to reuse.

At this point, the repository should be well-aligned with
Atmel Start.  Branch xKM_converser_d51-gg- or similar
lineage was the starting point, today.

Atmel Start official main.c .. to be rescinded in
the very next commit:

	modified:   main.c